2021/12/7

+++ src/main/java/com/example/bdz/service/impl/GwRoleServiceImpl.java
@@ -4,11 +4,8 @@ import cn.hutool.core.util.StrUtil;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.example.bdz.common.lang.Const;
 import com.example.bdz.common.lang.Result;
-import com.example.bdz.pojo.GwEquip;
-import com.example.bdz.pojo.GwRole;
+import com.example.bdz.pojo.*;
 import com.example.bdz.mapper.GwRoleMapper;
-import com.example.bdz.pojo.GwRoleMenu;
-import com.example.bdz.pojo.GwUserRole;
 import com.example.bdz.service.GwRoleMenuService;
 import com.example.bdz.service.GwRoleService;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
@@ -65,15 +62,39 @@ public class GwRoleServiceImpl extends ServiceImpl<GwRoleMapper, GwRole> impleme
     //添加角色
     @Override
     public Result addRole(GwRole gwRole) {
+        GwRole r=getRoleByName(gwRole.getName());
+        if(r!=null){
+            return Result.fail("角色已存在");
+        }
+        GwRole r1=getRoleByCode(gwRole.getCode());
+        if(r1!=null){
+            return Result.fail("角色编码已存在");
+        }
         gwRole.setStatus(Const.STATUS_ON);
         save(gwRole);
         return Result.success(gwRole);
+    }
+
+    private GwRole getRoleByCode(String code) {
+        return getOne(new QueryWrapper<GwRole>().eq("code",code));
+    }
+
+    private GwRole getRoleByName(String name) {
+        return getOne(new QueryWrapper<GwRole>().eq("name",name));
     }
 
     //编辑角色
     @Override
     public Result updateRole(Integer id, GwRole gwRole) {
         Assert.notNull(getById(id),"找不到该角色");
+        GwRole r=getRoleByName(gwRole.getName());
+        if(r!=null){
+            return Result.fail("角色已存在");
+        }
+        GwRole r1=getRoleByCode(gwRole.getCode());
+        if(r1!=null){
+            return Result.fail("角色编码已存在");
+        }
         gwRole.setId(id);
         updateById(gwRole);
         //清除缓存
@@ -87,6 +108,9 @@ public class GwRoleServiceImpl extends ServiceImpl<GwRoleMapper, GwRole> impleme
     public Result deleteRole(Integer id) {
         //删除缓存（先删除缓存，不然删除中间表之后会找不到对应的id）
         gwUserService.clearUserAuthorityInfoByRoleId(id);
+        if(id==1||id==2){
+            return Result.fail("该角色无法删除");
+        }
         //删除中间表
         gwUserRoleService.remove(new QueryWrapper<GwUserRole>().eq("role_id",id));
         gwRoleMenuService.remove(new QueryWrapper<GwRoleMenu>().eq("role_id",id));
